pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecosignallab"
version = "0.1.0"
description = "ecoSignalLab (esl): calibration-aware, multi-channel environmental and architectural acoustics toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "ecoSignalLab contributors" }
]
keywords = [
  "acoustics",
  "audio",
  "dsp",
  "ecoacoustics",
  "architectural-acoustics",
  "machine-learning",
  "signal-processing",
  "multichannel",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Sound/Audio :: Analysis",
  "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "soundfile>=0.12",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "h5py>=3.10",
  "requests>=2.31",
]

[project.optional-dependencies]
ml = [
  "torch>=2.0",
  "datasets>=2.16",
  "huggingface-hub>=0.20",
  "scikit-learn>=1.4",
]
io = [
  "pyarrow>=14.0",
]
plot = [
  "plotly>=5.18",
]
mat = [
  "scipy>=1.10",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.5",
  "mypy>=1.9",
]

[project.scripts]
esl = "esl.cli.main:main"

[project.urls]
Homepage = "https://github.com/example/ecoSignalLab"
Documentation = "https://github.com/example/ecoSignalLab"
Repository = "https://github.com/example/ecoSignalLab"
Issues = "https://github.com/example/ecoSignalLab/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = "src"
